B2: collapse Sql/Alembic renderer dispatch into single Renderers.render
Adding a MigrationOp case used to require edits in 3 files (the enum,
SqlRenderer.forward, AlembicRenderer.forward) — exactly the cascade pain
M7.6 hit when introducing AddTrigger/DropTrigger.

Introduce migration.Renderers.render(op): Rendered (sql, alembic). One
exhaustive pattern match emits both target outputs side-by-side. SqlRenderer
and AlembicRenderer become 4-line wrappers that pluck the relevant field
and add upgrade/downgrade scaffolding.

Net effect: future MigrationOp cases need ONE arm in Renderers.render plus
one inverse arm in MigrationOp. SchemaDiff still owns op emission (intrinsic
to its job).

Tests: all 131 codegen tests pass.
